pyproject.toml
[project]
name = "careers-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.5",
    "playwright>=1.40.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tqdm>=4.67.1",
]
